#479: minimum viable solution, i.e. it solved the bug - stylesets are displayed and the error is no longer observed in the console - but the font weight plugin does *not* work.
There are, however, other bugs on this, e.g. #459, #462 and #503, so the plugin code definitely needs to be revisited. We should consider disabling the plugin for now.

The submodule commit id in _scripler_ will be updated to the commit id of this commit when/if this pull request is approved.